requirements.txt
Flask==1.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
SQLAlchemy==1.3.11
